reorganize and enhance crypto utility functions
