+++ src/app/network/NetworkFacade.ts
@@ -9,10 +9,10 @@ import envService from 'app/core/services/env.service';
 import { buildProgressStream } from 'app/core/services/stream.service';
 import { queue, QueueObject } from 'async';
 import { WORKER_MESSAGE_STATES } from '../../WebWorker';
-import { generateFileKey } from '../drive/services/network.service';
+
 import { waitForContinueUploadSignal } from '../drive/services/worker.service/uploadWorkerUtils';
 import { TaskStatus } from '../tasks/types';
-import { encryptStreamInParts, getEncryptedFile, processEveryFileBlobReturnHash } from './crypto';
+import { encryptStreamInParts, generateFileKey, getEncryptedFile, processEveryFileBlobReturnHash } from './crypto';
 import { DownloadProgressCallback, getDecryptedStream } from './download';
 import { uploadFileBlob, UploadProgressCallback } from './upload';
